Create fctables - to stay logged in
Below we have tried to stay logged in with the help of passing cookies.

a) Following the steps we discussed in class and write code that automatically logs into the website fctables.com Links to an external site..

b) Verify that you have successfully logged in:  use the cookies you received during log in and write code to access https://www.fctables.com/tipster/my_bets/ Links to an external site..  Check whether the word “Wolfsburg” appears on the page.  Don’t look for your username to confirm that you are logged in (it won’t work) and use this page’s content instead.